% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dshield-package.R
\docType{package}
\name{dshield}
\alias{dshield}
\alias{dshield-package}
\title{Query 'SANS' 'DShield' 'API'}
\description{
The 'DShield' project provides early warning attack data and
host/'IP' metadata to the information security community. Tools are provided
to query various 'DShield' 'API' \url{https://isc.sans.edu/api/} endpoints.
}
\author{
Bob Rudis (bob@rud.is)
}
