% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port-history.R
\name{port_history}
\alias{port_history}
\title{Historical activity for a given port}
\usage{
port_history(port, start, end = Sys.Date())
}
\arguments{
\item{port}{TCP/UDP port}

\item{start, end}{date range to query within (ISO character or \code{Date} object).
\code{end} defaults to today.}
}
\description{
Historical activity for a given port
}
\examples{
port_history(5555, "2018-06-01")
}
